feat: update authorization logic in TokensController and UsersController; remove old policies and add new UsersPolicy and CustomerPolicy
