--- app/policies/admin_user_policy.rb
@@ -1,19 +1,0 @@
-class AdminUserPolicy < ApplicationPolicy
-  def create?
-    admin_and_confirmed?
-  end
-
-  def update?
-    admin_and_confirmed?
-  end
-
-  def black_list?
-    admin_and_confirmed?
-  end
-
-  private
-
-  def admin_and_confirmed?
-    user.admin? && user.confirmed?
-  end
-end
