--- app/policies/user_policy.rb
@@ -1,5 +1,0 @@
-class UserPolicy < ApplicationPolicy
-  def verify?
-    user.id == record.id
-  end
-end
