+++ app/policies/v1/users/verifications_policy.rb
@@ -1,0 +1,9 @@
+module V1
+  module Users
+    class VerificationsPolicy < ApplicationPolicy
+      def create?
+        user.id == record.id
+      end
+    end
+  end
+end
